pyproject.toml
[tool.poetry]
name = "text-recog"
version = "0.1.0"
description = ""
authors = ["ratuozzo <ratuozzo@gmail.com>"]
readme = "README.md"
packages = [{include = "text_recog"}]

[tool.poetry.dependencies]
python = "^3.11"
azure-cognitiveservices-speech = "^1.29.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
